WRITE-UP/MY PROJECT/DeepNeural/requirements.txt
streamlit
google-generativeai
python-dotenv
venv_name\Scripts\activate